renames file to prevent overwriting
